Move response widgets to separate components
