+++ src/components/surveyForm/widgets/TextWidget.jsx
@@ -1,0 +1,33 @@
+import React from 'react';
+
+import PropTypes from '../../../utils/PropTypes';
+
+
+export default class TextWidget extends React.Component {
+    static propTypes = {
+        name: PropTypes.string.isRequired,
+        question: PropTypes.map.isRequired,
+    };
+
+    render() {
+        let name = this.props.name;
+        let question = this.props.question;
+
+        if (question.getIn(['response_config', 'multiline'])) {
+            var widget = (
+                <textarea name={ name + '.text' }/>
+            );
+        }
+        else {
+            var widget = (
+                <input name={ name + '.text' }/>
+            );
+        }
+
+        return (
+            <div className="TextWidget">
+                { widget }
+            </div>
+        );
+    }
+}
